+++ tst/com/amazon/corretto/crypto/provider/test/EvpKeyAgreementTest.java
@@ -402,13 +402,18 @@ public class EvpKeyAgreementTest {
         assertThrows(InvalidKeyException.class,
                 () -> agree.init(new SecretKeySpec("YellowSubmarine".getBytes(StandardCharsets.UTF_8), "AES")));
 
+        assertThrows(InvalidKeyException.class, () -> agree.init(null));
+
         assertThrows(InvalidAlgorithmParameterException.class,
                 () -> agree.init(pairs[0].getPrivate(), new IvParameterSpec(new byte[0])));
 
         agree.init(pairs[0].getPrivate(), (AlgorithmParameterSpec) null);
 
-        assertThrows(IllegalStateException.class, "Only single phase agreement is supported",
-                () -> agree.doPhase(pairs[0].getPublic(), false));
+        // This test doesn't apply to DH
+        if (!algorithm.equals("DH")) {
+            assertThrows(IllegalStateException.class, "Only single phase agreement is supported",
+                    () -> agree.doPhase(pairs[0].getPublic(), false));
+        }
         assertThrows(IllegalStateException.class, "KeyAgreement has not been completed", () -> agree.generateSecret());
 
         assertThrows(InvalidKeyException.class,